move TritonLinks to utils
